+++ TwinTech/TwinsanityInterchange/Implementations/PS2/Items/Graphics/PS2AnyTexture.cs
@@ -216,8 +216,8 @@ namespace Twinsanity.TwinsanityInterchange.Implementations.PS2.Items.Graphics
                 MipLevels = 1;
             }
             //this is probably not bytes but whatever
-            UnkBytes2[5] = UnkBytes3[0] = (width == 256) ? 0 : (byte)Math.Min(width, height);
-            UnkBytes2[6] = UnkBytes3[1] = (width == 256) ? 2 : 0;
+            UnkBytes2[5] = UnkBytes3[0] = (Byte)((width == 256) ? 0 : (byte)Math.Min(width, height));
+            UnkBytes2[6] = UnkBytes3[1] = (Byte)((width == 256) ? 2 : 0);
            
             GIFTag headerTag = new GIFTag();
             headerTag.REGS = new REGSEnum[16];
